[AMQCPP-56] - make buffer lengths use size_t for 64-bit operating systems
git-svn-id: https://svn.apache.org/repos/asf/incubator/activemq/activemq-cpp/trunk@500349 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-cpp/src/main/activemq/connector/openwire/marshal/PrimitiveMapMarshaller.cpp
@@ -48,7 +48,7 @@ void PrimitiveMapMarshaller::marshal( const util::PrimitiveMap* map,
         }
         else
         {
-            dataOut.writeInt( map->size() );
+            dataOut.writeInt( (int)map->size() );
             
             std::vector<std::string> keys = map->getKeys();
             std::vector<std::string>::const_iterator iter = keys.begin();
@@ -150,7 +150,7 @@ void PrimitiveMapMarshaller::marshalPrimitive( io::DataOutputStream& dataOut,
             
             std::vector<unsigned char> data = value.getByteArray();
             
-            dataOut.writeInt( data.size() );
+            dataOut.writeInt( (int)data.size() );
             dataOut.write( data );
         }
         else if( value.getValueType() == PrimitiveMap::STRING_TYPE )
